+++ sdk/src/client/storage/mod.rs
@@ -18,15 +18,15 @@ pub trait StorageProvider {
     type Error;
 
     /// Get a value out of the database.
-    async fn get(&mut self, k: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
+    async fn get(&self, k: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
 
     /// Insert a value into the database.
     ///
     /// If there exists a record under the same key as `k`, it will be replaced by the new value (`v`) and returned.
-    async fn insert(&mut self, k: &[u8], v: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
+    async fn insert(&self, k: &[u8], v: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
 
     /// Delete a value from the database.
     ///
     /// The deleted value is returned.
-    async fn delete(&mut self, k: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
+    async fn delete(&self, k: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
 }